Move types to separate file. Add get individual game function
